requirements.txt
python-dotenv>=0.10.1
paho-mqtt>=1.4.0
